+++ coralogix/data_source_coralogix_api_key.go
@@ -27,7 +27,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 var _ datasource.DataSourceWithConfigure = &ApiKeyDataSource{}
@@ -88,7 +87,7 @@ func (d *ApiKeyDataSource) Read(ctx context.Context, req datasource.ReadRequest,
 	getApiKeyResponse, err := d.client.Get(ctx, getApiKey)
 	if err != nil {
 		log.Printf("[ERROR] Received error: %s", err.Error())
-		if status.Code(err) == codes.NotFound {
+		if cxsdk.Code(err) == codes.NotFound {
 			resp.Diagnostics.AddWarning(err.Error(),
 				fmt.Sprintf("API Keys %q is in state, but no longer exists in Coralogix backend", id),
 			)